api/requirements.txt
google-adk==1.10.0